requirements.txt
# psycopg2
boto3
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.7
distlib==0.3.0
dj-database-url==0.5.0
Django==2.2
django-adminactions==1.9
djangorestframework==3.13.1
django-ads==1.1.1
django-allauth==0.39.1
django-appconf==1.0.4
django-bootstrap-form==3.4
django-braces==1.14.0
django-ckeditor==5.9.0
django-classy-tags==3.0.0
django-cors-headers==3.11.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==2.2.0
django-etc==1.2.0
django-extra-views==0.11.0
django-filter==2.2.0
django-fontawesome-5==1.0.18
django-haystack==2.8.1
django-heroku==0.3.1
django-hitcount==1.3.3
django-inline-svg==0.1.1
django-js-asset==1.2.2
django-js-reverse==0.9.1
django-modelcluster==5.0.1
django-sass==1.0.0
gunicorn==19.9.0
django_environ==0.9.0
asgiref==3.1.2
whitenoise
# svg
django-storages==1.10.0
django-cms==3.7.0
django-sekizai==3.0.1
django-model-utils==4.1.0
pytest==7.2.0
pytest-django==4.5.0
"django-haystack[elasticsearch]"


